+++ docs/transform_index.py
@@ -115,6 +115,7 @@ body {
     line-height: 30px;
     display: inline-block;
     margin-top: 10px;
+    color: white;
 }
 #sidebar {
     height: 94vh;
@@ -123,7 +124,7 @@ body {
     top: 59px;
     left: 0;
     background-color: #f4f4f4;
-	overflow-y: auto;
+    overflow-y: auto;
     z-index: 999;
 }
 #sidebar a {
@@ -196,9 +197,9 @@ function offsetAnchor() {
 
 </script>
 
-<div id='top-menu'>
-    <img src="images/isanxot/isanxot.png" alt="Logo" id="logo">
-    <div id="title">iSanXoT</div>
+<div id="top-menu">
+    <a href="https://cnic-proteomics.github.io/iSanXoT"><img src="images/isanxot/isanxot.png" alt="Logo" id="logo"></a>
+    <a href="https://cnic-proteomics.github.io/iSanXoT" id="title-link"><div id="title">iSanXoT</div></a>
 </div>
 <div id='sidebar'>
 
@@ -265,6 +266,7 @@ function offsetAnchor() {
       <li><a href="#_Workflow_2:_Step-by-step">Wkf 2: Step-by-step quantification and sample combination</a></li>
       <li><a href="#_Workflow_3:_Quantification">Wkf 3: PTMs quantification</a></li>
       <li><a href="#_Workflow_4:_Label-free">Wkf 4: Label-free quantification</a></li>
+      <li><a href="#_Toc187758918">Wkf 5: PTM-compass</a></li>
     </ul>
   </li>
   <li><a href="#_Toc152521038">Importing a workflow template</a></li>